Use base handler for email verification job
